Fixed problem with Bean creation. Application starts without errors. Needs testing

+++ outbound-repository-jpa/src/main/java/org/windat/jpa/repository/LobbySpringDataRepository.java
@@ -3,5 +3,5 @@ package org.windat.jpa.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.windat.domain.entity.Lobby;
 
-public interface LobbySpringDataRepository extends JpaRepository<Lobby, Long> {
+public interface LobbySpringDataRepository extends JpaRepository<Lobby, Integer> {
 }
